requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
attrs==23.2.0
cca-zoo==2.6.0
colorama==0.4.6
colorlog==6.8.2
contourpy==1.1.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lightning==2.2.4
lightning-utilities==0.11.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl==2021.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
opt-einsum==3.3.0
optuna==3.6.1
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pretty-errors==1.2.25
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.4
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.30
sympy==1.12
tbb==2021.12.0
tensorly==0.8.1
threadpoolctl==3.5.0
torch==2.3.0+cu118
torchaudio==2.3.0
torchmetrics==1.4.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
yarl==1.9.4
zipp==3.18.1
